Integrate JWT Auth into API
partially fixes #2063

This will allow us to add permissions checks to our API calls. Currently
it does not require a valid JWT, but if one exists will store the "sub"
field such that the user's id can be checked against our user table to
establish permissions.

In the user profile api calls we will check user permission, but not yet
fail (just log) if the constraint is not met.