Demonstration of getting user_id from auth token
This will require more work to actually function, but this code
automatically attempts to parse the auth0 bearer token from any request
made to the api and then makes the user_id available as part of the
application context to the route implementations.

This currently does not work. Most likely we have to configure an API
in auth0 associated with the SPA, but there may be other ways around
that.

Currently the proximal issue is that the token has "alg" set to "dir"
which is not an accepted algorithm by the aulib code used to parse the
token. See issue log for more information.