Prevent concurrent deploys.
Fixes #2423

Running multiple deploys concurrently will result in race conditions. We believe this is currently causing the gcloud app deploy command to be confused about which previous version or versions to stop, but it could cause other issues as well.

This change uses a concurrency group to only permit a single deploy to execute at a time.